Fix blob delete bug (#177)
* Fix blob delete bug

* Cleanup

+++ src/main/java/uk/gov/hmcts/reform/bulkscanprocessor/tasks/BlobProcessorTask.java
@@ -14,6 +14,7 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.bulkscanprocessor.entity.Envelope;
+import uk.gov.hmcts.reform.bulkscanprocessor.entity.Status;
 import uk.gov.hmcts.reform.bulkscanprocessor.services.wrapper.ErrorHandlingWrapper;
 import uk.gov.hmcts.reform.bulkscanprocessor.tasks.processor.DocumentProcessor;
 import uk.gov.hmcts.reform.bulkscanprocessor.tasks.processor.EnvelopeProcessor;
@@ -77,10 +78,10 @@ public class BlobProcessorTask extends Processor {
         log.info("Processing zip file {}", zipFilename);
 
         CloudBlockBlob cloudBlockBlob = container.getBlockBlobReference(zipFilename);
-        Envelope failedDeleteEnvelope =
-            envelopeProcessor.getIfFailToDeleteBlobBefore(container.getName(), zipFilename);
-        if (failedDeleteEnvelope != null) {
-            delete(cloudBlockBlob, failedDeleteEnvelope);
+        Envelope existingEnvelope =
+            envelopeProcessor.getEnvelopeByFileAndContainer(container.getName(), zipFilename);
+        if (existingEnvelope != null) {
+            deleteIfProcessed(cloudBlockBlob, existingEnvelope);
             return;
         }
 
@@ -103,10 +104,10 @@ public class BlobProcessorTask extends Processor {
             }
         }
     }
-
-    private void delete(CloudBlockBlob cloudBlockBlob, Envelope envelope) {
+    
+    private void deleteIfProcessed(CloudBlockBlob cloudBlockBlob, Envelope envelope) {
         try {
-            if (cloudBlockBlob != null) {
+            if (cloudBlockBlob != null && envelope.getStatus() == Status.PROCESSED) {
                 boolean deleted;
                 if (cloudBlockBlob.exists()) {
                     deleted = cloudBlockBlob.deleteIfExists();
